umbrella potential bugs at periodic boundaries fixed
Major updates:
Bug that caused ions to become trapped against the graphene sheet
was fixed by forcing the ion of interest to the center of the
periodic image before applying the potential. This should ensure
that the shortest path from the ion to the potential is never through
the graphene sheet, even if the ion is dilute.

There was a bug that caused the umbrella potential to set itself at
the periodic image boundary if the pore was split by the boundary.
This was solved by shifting the potential a box length in the dimension
along which the pore was split.

Minor updates:
the pmf conversion script reads how much to shift the potential each
time the window shifts from the output script

a couple of flags were added to the run script
    - trajFreq: changes how often the trajectory file gets written
    - gpuDevice: changes the GPU device :)