determining pore center uses minimum image distances
When equilibration positioned the pore such that it laid on the boundary
between periodic images, the old pore center code would calculate a "center"
that laid exactly between images because it averaged the positions of atoms
within the periodic box. This commit fixes that issue, ensuring that the closest
periodic distance to an atom is always used.